refactor: 각 객체가 DB의 bigint를 지원하도록 id 값을 Long Tpye으로 변

+++ src/main/java/kr/co/dbsg/api/api/event/repository/EventRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRepository extends JpaRepository<EventEntity, Integer>, JpaSpecificationExecutor<EventEntity> {
+public interface EventRepository extends JpaRepository<EventEntity, Long>, JpaSpecificationExecutor<EventEntity> {
 }
